+++ sdk/iot_hub/src/service/responses/device_identity_response.rs
@@ -48,3 +48,14 @@ impl std::convert::TryFrom<crate::service::CollectedResponse> for DeviceIdentity
         Ok(device_identity_response)
     }
 }
+
+/// Response of CreateOrUpdateDeviceIdentity
+pub type CreateOrUpdateDeviceIdentityResponse = DeviceIdentityResponse;
+
+impl CreateOrUpdateDeviceIdentityResponse {
+    pub(crate) async fn try_from(response: azure_core::Response) -> azure_core::Result<Self> {
+        let collected = azure_core::CollectedResponse::from_response(response).await?;
+        let body = collected.body();
+        Ok(serde_json::from_slice(body)?)
+    }
+}
