+++ sdk/iot_hub/src/service/responses/mod.rs
@@ -7,9 +7,9 @@ mod module_twin_response;
 mod query_response;
 
 pub use configuration_response::{ConfigurationResponse, MultipleConfigurationResponse};
-pub use device_identity_response::DeviceIdentityResponse;
+pub use device_identity_response::{CreateOrUpdateDeviceIdentityResponse, DeviceIdentityResponse};
 pub use device_twin_response::DeviceTwinResponse;
 pub use invoke_method_response::InvokeMethodResponse;
-pub use module_identity_response::ModuleIdentityResponse;
+pub use module_identity_response::{CreateOrUpdateModuleIdentityResponse, ModuleIdentityResponse};
 pub use module_twin_response::ModuleTwinResponse;
 pub use query_response::QueryResponse;
